rename source dir to ts and ignore dist
