+++ .gitignore
@@ -1,2 +1,3 @@
 release_builds/
 node_modules/
+dist/
