+++ app/ts/common/dnsLookup.ts
