+++ app/ts/common/fontawesome/all.js
