+++ app/ts/common/fontawesome/fontawesome.js
