+++ app/ts/common/fontawesome/regular.js
