+++ app/ts/common/fontawesome/solid.js
