+++ app/ts/common/fontawesome/v4-shims.js
