+++ app/ts/common/lineHelper.ts
