+++ app/ts/common/parseRawData.ts
