+++ app/ts/common/resetObject.ts
