+++ app/ts/common/settings.ts
