+++ app/ts/common/stringformat.ts
