+++ app/ts/main/bwa/analyser.ts
