+++ app/ts/renderer/bwa/analyser.ts
