+++ tsconfig.json
@@ -6,14 +6,14 @@
     "moduleResolution": "node",
     "allowJs": true,
     "noEmit": false,
-    "rootDir": "app/js",
-    "outDir": "dist/app/js",
+    "rootDir": "app/ts",
+    "outDir": "dist/app/ts",
     "esModuleInterop": true,
     "strict": true,
     "skipLibCheck": true,
     "types": ["node", "jest"],
     "typeRoots": ["./types", "./node_modules/@types"]
   },
-  "include": ["app/js/**/*.ts", "**/*.d.ts"],
+  "include": ["app/ts/**/*.ts", "**/*.d.ts"],
   "exclude": ["test"]
 }
